+++ app/src/main/java/com/poli/prevendasomie/data/remote/responses/pedidos/Irrf.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Irrf(
     @SerializedName("aliq_irrf")
-    val aliqIrrf: Int?,
+    val aliqIrrf: Double?,
     @SerializedName("valor_irrf")
-    val valorIrrf: Int?
+    val valorIrrf: Double?
 )
