+++ app/src/main/java/com/poli/prevendasomie/data/remote/responses/pedidos/PisSt.kt
@@ -4,19 +4,19 @@ import com.google.gson.annotations.SerializedName
 
 data class PisSt(
     @SerializedName("aliq_pis_st")
-    val aliqPisSt: Int?,
+    val aliqPisSt: Double?,
     @SerializedName("base_pis_st")
-    val basePisSt: Int?,
+    val basePisSt: Double?,
     @SerializedName("cod_sit_trib_pis_st")
     val codSitTribPisSt: String?,
     @SerializedName("margem_pis_st")
-    val margemPisSt: Int?,
+    val margemPisSt: Double?,
     @SerializedName("qtde_unid_trib_pis_st")
     val qtdeUnidTribPisSt: Int?,
     @SerializedName("tipo_calculo_pis_st")
     val tipoCalculoPisSt: String?,
     @SerializedName("valor_pis_st")
-    val valorPisSt: Int?,
+    val valorPisSt: Double?,
     @SerializedName("valor_unid_trib_pis_st")
-    val valorUnidTribPisSt: Int?
+    val valorUnidTribPisSt: Double?
 )
