+++ app/src/main/java/com/poli/prevendasomie/domain/model/pedidos/Parcela.kt
@@ -6,7 +6,7 @@ data class Parcela(
     @SerializedName("categoria_adiantamento")
     val categoriaAdiantamento: String?,
     @SerializedName("conta_corrente_adiantamento")
-    val contaCorrenteAdiantamento: Int?,
+    val contaCorrenteAdiantamento: Long?,
     @SerializedName("data_vencimento")
     val dataVencimento: String?,
     @SerializedName("meio_pagamento")
